Fixed execution prem, bench should block at prio 2

+++ src/hypervisor.c
@@ -1,6 +1,6 @@
 #include <hypervisor.h>
 
-uint64_t hypercall(enum hypervisor_actions action, int arg0, int arg1, int arg2) {
+uint64_t hypercall(enum hypervisor_actions action, uint64_t arg0, uint64_t arg1, uint64_t arg2) {
     uint64_t hypercall_id = HYPERCALL_BASE_VALUE | action;
     uint64_t ret;
